+++ src/index.ts
@@ -1,5 +1,6 @@
 import { api } from './api';
 import { AccountService } from './api/AccountService';
+import { AuthenticationService } from './api/AuthenticationService';
 import { CertificationsService } from './api/CertificationsService';
 import { ChangesService } from './api/ChangesService';
 import { CollectionsService } from './api/CollectionsService';
@@ -75,6 +76,8 @@ export default class TMDB {
 
 	public collections: CollectionsService;
 
+	public authentication: AuthenticationService;
+
 	/**
 	 * Creates an instance of the TMDB class.
 	 * @param apiKey - The API key used for authenticating requests.
@@ -93,5 +96,6 @@ export default class TMDB {
 		this.certifications = new CertificationsService(this.apiInstance);
 		this.changes = new ChangesService(this.apiInstance);
 		this.collections = new CollectionsService(this.apiInstance);
+		this.authentication = new AuthenticationService(this.apiInstance);
 	}
 }